Fix sql migrations, change type from BIGINT to NUMERIC for decimal numbers

+++ src/main/resources/db/migration/V4__gift-card.sql
@@ -4,7 +4,7 @@ CREATE TABLE gift_card (
     owner_id CHAR(36),
     store_id CHAR(36),
     gift_card_status VARCHAR(7) NOT NULL CHECK (gift_card_status IN ('PENDING', 'ACTIVE', 'EXPIRED', 'USED_UP')),
-    balance BIGINT NOT NULL,
+    balance NUMERIC NOT NULL,
     count_of_uses SMALLINT NOT NULL,
     max_count_of_uses SMALLINT NOT NULL,
     secret_key VARCHAR NOT NULL,
